feat(USM) implement debug switch
