+++ Makefile
@@ -21,16 +21,16 @@ build: elf
 	cd ./cmd;CGO_ENABLED=0 GOOS=$(GOOS) GOARCH=$(GOARCH)   go build -gcflags "all=-N -l" -o nfs-trace
 
 dlv:  build
-	dlv --headless --listen=:2345 --api-version=2 exec ./cmd/nfs-trace-linux-amd64 -- -filter-struct=$(FILTER_STRUCT) -filter-func="^nfs.*" -all-kmods=true -output-metrics
+	dlv --headless --listen=:2345 --api-version=2 exec ./cmd/nfs-trace -- --filter-struct=$(FILTER_STRUCT) --filter-func="^nfs.*"  --enable-debug=true --all-kmods=true
 
 run:  build
-	./cmd/nfs-trace-linux-amd64 -filter-struct=$(FILTER_STRUCT) -filter-func="^(vfs_|nfs_).*" -all-kmods=true
+	./cmd/nfs-trace --filter-struct=$(FILTER_STRUCT) --filter-func="^(vfs_|nfs_).*" --enable-debug=true --all-kmods=true 
 
 skip:  build
-	./cmd/nfs-trace-linux-amd64 -filter-struct=$(FILTER_STRUCT) -skip-attach=true -all-kmods=true -filter-func="^nfs.*"
+	./cmd/nfs-trace --filter-struct=$(FILTER_STRUCT) --skip-attach=true --all-kmods=true --filter-func="^nfs.*"
 
 funcs:  build
-	./cmd/nfs-trace-linux-amd64 -filter-struct=kiocb -all-kmods=true -filter-func="^(vfs_|nfs_).*" -add-funcs="nfs_file_direct_read:1,nfs_file_direct_write:1,nfs_swap_rw:1,nfs_file_read:1,nfs_file_write:1"
+	./cmd/nfs-trace --filter-struct=kiocb --all-kmods=true --filter-func="^(vfs_|nfs_).*" --add-funcs="nfs_file_direct_read:1,nfs_file_direct_write:1,nfs_swap_rw:1,nfs_file_read:1,nfs_file_write:1"
 
 elf:
 	TARGET_GOARCH=$(TARGET_GOARCH) FILTER_STRUCT=$(FILTER_STRUCT) $(GO_GENERATE)
