+++ bpf/kprobe_pwru.c
@@ -10,6 +10,14 @@
 #include "bpf/bpf_endian.h"
 #include "bpf/bpf_ipv6.h"
 
+struct config
+{
+    u8 debug_log;
+} __attribute__((packed));
+
+static volatile const struct config CFG;
+#define cfg (&CFG)
+
 struct raw_metrics
 {
     u64 read_count;
@@ -329,7 +337,10 @@ int nfs_init_write(struct nfs_init_fields *ctx)
     u32 tid = (u32)bpf_get_current_pid_tgid();
     u64 timestamp = bpf_ktime_get_ns();
 
-    // bpf_printk("nfs_init_write: %llu, pid: %u, tid: %u\n", timestamp, pid, tid);
+    if (cfg->debug_log)
+    {
+        bpf_printk("nfs_init_write: %llu, pid: %u, tid: %u\n", timestamp, pid, tid);
+    }
 
     bpf_map_update_elem(&link_begin, &pid, &timestamp, BPF_ANY);
 
@@ -343,7 +354,10 @@ int rpc_task_begin(struct rpc_task_state *ctx)
     u32 tid = (u32)bpf_get_current_pid_tgid();
     u64 rpc_task_id = (u64)ctx->task_id;
 
-    // bpf_printk("rpc_task_begin: %llu, pid: %u, tid: %u\n", rpc_task_id, pid, tid);
+    if (cfg->debug_log)
+    {
+        bpf_printk("rpc_task_begin: %llu, pid: %u, tid: %u\n", rpc_task_id, pid, tid);
+    }
 
     u64 *timestamp = bpf_map_lookup_elem(&link_begin, &pid);
     if (timestamp)
@@ -363,7 +377,10 @@ int rpc_task_done(struct rpc_task_state *ctx)
 {
     u64 rpc_task_id = (u64)ctx->task_id;
 
-    // bpf_printk("rpc_task_done: %llu\n", rpc_task_id);
+    if (cfg->debug_log)
+    {
+        bpf_printk("rpc_task_done: %llu\n", rpc_task_id);
+    }
 
     struct rpc_task_info *info = bpf_map_lookup_elem(&waiting_RPC, &rpc_task_id);
     if (info)
@@ -396,7 +413,10 @@ int rpc_execute(struct pt_regs *regs)
     u32 tid = (u32)bpf_get_current_pid_tgid();
     u64 rpc_task_id = BPF_CORE_READ(task, tk_pid);
 
-    // bpf_printk("rpc_execute: %llu, pid: %u, tid: %u\n", rpc_task_id, pid, tid);
+    if (cfg->debug_log)
+    {
+        bpf_printk("rpc_execute: %llu, pid: %u, tid: %u\n", rpc_task_id, pid, tid);
+    }
 
     u64 *timestamp = bpf_map_lookup_elem(&link_begin, &pid);
     if (timestamp)
@@ -419,7 +439,10 @@ int rpc_exit_task(struct pt_regs *regs)
 
     u64 rpc_task_id = BPF_CORE_READ(task, tk_pid);
 
-    // bpf_printk("rpc_exit_task: %llu\n", rpc_task_id);
+    if (cfg->debug_log)
+    {
+        bpf_printk("rpc_exit_task: %llu\n", rpc_task_id);
+    }
 
     struct rpc_task_info *info = bpf_map_lookup_elem(&waiting_RPC, &rpc_task_id);
     if (info)
@@ -490,8 +513,11 @@ int kb_nfs_read_d(struct pt_regs *regs)
     // 更新 io_metrics map
     bpf_map_update_elem(&io_metrics, &key, metrics, BPF_ANY);
 
-    // bpf_printk("Read - dev: %llu, file: %llu, bytes: %u, count: %d, total_bytes: %d, latency: %d\n",
-    //            dev, fileid, res_count, metrics->read_count, metrics->read_size, metrics->read_lat);
+    if (cfg->debug_log)
+    {
+        bpf_printk("Read - dev: %llu, file: %llu, bytes: %u, count: %d, total_bytes: %d, latency: %d\n",
+                   dev, fileid, res_count, metrics->read_count, metrics->read_size, metrics->read_lat);
+    }
 
     return 0;
 }
@@ -544,8 +570,11 @@ int kb_nfs_write_d(struct pt_regs *regs)
     // 更新 io_metrics map
     bpf_map_update_elem(&io_metrics, &key, metrics, BPF_ANY);
 
-    // bpf_printk("Write - dev: %llu, file: %llu, bytes: %u, count: %d, total_bytes: %d, latency: %d\n",
-    //            dev, fileid, res_count, metrics->write_count, metrics->write_size, metrics->write_lat);
+    if (cfg->debug_log)
+    {
+        bpf_printk("Write - dev: %llu, file: %llu, bytes: %u, count: %d, total_bytes: %d, latency: %d\n",
+                   dev, fileid, res_count, metrics->write_count, metrics->write_size, metrics->write_lat);
+    }
 
     return 0;
 }
